fix parsing of atomic composition in excited species A(*)
